+++ packages/sdk/src/base/index.ts
@@ -1,5 +1,4 @@
 export * from './storage';
-export * from './transport';
 
 export * from './auth';
 export * from './directus';
