+++ packages/sdk/src/handlers/activity.ts
@@ -3,7 +3,7 @@
  */
 
 import { ItemsHandler } from '../base/items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { ActivityType, DefaultType } from '../types';
 import { CommentsHandler } from './comments';
 
@@ -12,7 +12,7 @@ export type ActivityItem<T = DefaultType> = ActivityType & T;
 export class ActivityHandler<T = DefaultType> extends ItemsHandler<ActivityItem<T>> {
 	private _comments: CommentsHandler<T>;
 
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		super('directus_activity', transport);
 		this._comments = new CommentsHandler(this.transport);
 	}
