+++ packages/sdk/src/handlers/extensions.ts
@@ -2,62 +2,66 @@
  * Settings handler
  */
 
-import { ITransport, TransportOptions, TransportResponse } from '../transport';
+/** !!! Rework this one !!! */
 
-export class ExtensionEndpoint implements ITransport {
-	private name: string;
-	private transport: ITransport;
+// import { Transport, TransportRequestOptions, TransportResponse } from '../transport';
 
-	constructor(transport: ITransport, name: string) {
-		this.name = name;
-		this.transport = transport;
-	}
+export {};
 
-	private endpoint(path: string) {
-		if (path.startsWith('/')) {
-			path = path.substr(1);
-		}
+// export class ExtensionEndpoint extends Transport {
+// 	private name: string;
+// 	private transport: Transport;
 
-		return `/custom/${this.name}/${path}`;
-	}
+// 	constructor(transport: Transport, name: string) {
+// 		this.name = name;
+// 		this.transport = transport;
+// 	}
 
-	get<T = any>(path: string, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.get(this.endpoint(path), options);
-	}
+// 	private endpoint(path: string) {
+// 		if (path.startsWith('/')) {
+// 			path = path.substr(1);
+// 		}
 
-	head<T = any>(path: string, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.head(this.endpoint(path), options);
-	}
+// 		return `/custom/${this.name}/${path}`;
+// 	}
 
-	options<T = any>(path: string, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.options(this.endpoint(path), options);
-	}
+// 	get<T = any>(path: string, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.get(this.endpoint(path), options);
+// 	}
 
-	delete<T = any, P = any>(path: string, data?: P, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.delete(this.endpoint(path), data, options);
-	}
+// 	head<T = any>(path: string, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.head(this.endpoint(path), options);
+// 	}
 
-	post<T = any, P = any>(path: string, data?: P, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.post(this.endpoint(path), data, options);
-	}
+// 	options<T = any>(path: string, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.options(this.endpoint(path), options);
+// 	}
 
-	put<T = any, P = any>(path: string, data?: P, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.put(this.endpoint(path), data, options);
-	}
+// 	delete<T = any, P = any>(path: string, data?: P, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.delete(this.endpoint(path), data, options);
+// 	}
 
-	patch<T = any, P = any>(path: string, data?: P, options?: TransportOptions): Promise<TransportResponse<T>> {
-		return this.transport.patch(this.endpoint(path), data, options);
-	}
-}
+// 	post<T = any, P = any>(path: string, data?: P, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.post(this.endpoint(path), data, options);
+// 	}
 
-export class ExtensionHandler {
-	private transport: ITransport;
+// 	put<T = any, P = any>(path: string, data?: P, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.put(this.endpoint(path), data, options);
+// 	}
 
-	constructor(transport: ITransport) {
-		this.transport = transport;
-	}
+// 	patch<T = any, P = any>(path: string, data?: P, options?: TransportRequestOptions): Promise<TransportResponse<T>> {
+// 		return this.transport.patch(this.endpoint(path), data, options);
+// 	}
+// }
 
-	endpoint(name: string): ExtensionEndpoint {
-		return new ExtensionEndpoint(this.transport, name);
-	}
-}
+// export class ExtensionHandler {
+// 	private transport: Transport;
+
+// 	constructor(transport: Transport) {
+// 		this.transport = transport;
+// 	}
+
+// 	endpoint(name: string): ExtensionEndpoint {
+// 		return new ExtensionEndpoint(this.transport, name);
+// 	}
+// }
