+++ packages/sdk/src/handlers/folders.ts
@@ -3,13 +3,13 @@
  */
 
 import { ItemsHandler } from '../base/items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { FolderType, DefaultType } from '../types';
 
 export type FolderItem<T = DefaultType> = FolderType & T;
 
 export class FoldersHandler<T = DefaultType> extends ItemsHandler<FolderItem<T>> {
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		super('directus_folders', transport);
 	}
 }
