+++ packages/sdk/src/handlers/invites.ts
@@ -1,10 +1,10 @@
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { ID } from '../types';
 
 export class InvitesHandler {
-	private transport: ITransport;
+	private transport: Transport;
 
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		this.transport = transport;
 	}
