+++ packages/sdk/src/handlers/presets.ts
@@ -3,13 +3,13 @@
  */
 
 import { ItemsHandler } from '../base/items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { PresetType, DefaultType } from '../types';
 
 export type PresetItem<T = DefaultType> = PresetType & T;
 
 export class PresetsHandler<T = DefaultType> extends ItemsHandler<PresetItem<T>> {
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		super('directus_presets', transport);
 	}
 }
