+++ packages/sdk/src/handlers/relations.ts
@@ -2,14 +2,14 @@
  * Relations handler
  */
 import { ManyItems, OneItem, ItemInput, EmptyParamError, Item } from '../items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { RelationType, DefaultType, ID } from '../types';
 
 export type RelationItem<T = DefaultType> = RelationType & T;
 export class RelationsHandler<T extends Item = RelationItem> {
-	transport: ITransport;
+	transport: Transport;
 
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		this.transport = transport;
 	}
