+++ packages/sdk/src/handlers/revisions.ts
@@ -3,13 +3,13 @@
  */
 
 import { ItemsHandler } from '../base/items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { RevisionType, DefaultType } from '../types';
 
 export type RevisionItem<T = DefaultType> = RevisionType & T;
 
 export class RevisionsHandler<T = DefaultType> extends ItemsHandler<RevisionItem<T>> {
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		super('directus_revisions', transport);
 	}
 }
