+++ packages/sdk/src/handlers/roles.ts
@@ -3,13 +3,13 @@
  */
 
 import { ItemsHandler } from '../base/items';
-import { ITransport } from '../transport';
+import { Transport } from '../transport';
 import { RoleType, DefaultType } from '../types';
 
 export type RoleItem<T = DefaultType> = RoleType & T;
 
 export class RolesHandler<T = DefaultType> extends ItemsHandler<RoleItem<T>> {
-	constructor(transport: ITransport) {
+	constructor(transport: Transport) {
 		super('directus_roles', transport);
 	}
 }
